web/requirements.txt
Flask>=2.0.0
flask_restful
pymongo
bcrypt
spacy
numpy
tensorflow
requests
keras
pillow